backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
Pillow==10.3.0
python-multipart==0.0.9
boto3==1.34.115
